Reset all session options to package defaults

Restores all ksTFL session options (headers, footers, body text, styles, page settings, column width behavior, etc.) to their original package defaults. Useful at the start of a new reporting session or after experimenting with tfl_set_options().

Usage

tfl_reset_options()

Value

The default options list, returned invisibly.

See also

Examples

# Set some session defaults
tfl_set_options(
  add_header("Study ABC", "Phase II", "CONFIDENTIAL"),
  add_footer("Company", "Page {PAGE}", "2025")
)

# ... build tables ...

# Reset everything back to package defaults
tfl_reset_options()
tfl_get_options()  # Confirm reset
#> $doc_style_template
#> [1] "Default"
#> 
#> $page
#> NULL
#> 
#> $footnotePlace
#> [1] "repeated"
#> 
#> $isContinues
#> [1] FALSE
#> 
#> $continuousSection
#> [1] FALSE
#> 
#> $contentWidth
#> [1] "100%"
#> 
#> $figureWidth
#> [1] "6in"
#> 
#> $figureHeight
#> [1] "4in"
#> 
#> $figureDevice
#> [1] "cairo"
#> 
#> $figureScaleMode
#> [1] "fixed"
#> 
#> $missings
#> [1] ""
#> 
#> $autoColWidth
#> [1] TRUE
#> 
#> $minColWidth
#> [1] 0.5
#> 
#> $headers
#> list()
#> attr(,"class")
#> [1] "TFL_options"
#> 
#> $footers
#> list()
#> attr(,"class")
#> [1] "TFL_options"
#> 
#> $bodyText
#> $`__default_001`
#> $`__default_001`$text
#> [1] "No data to report"
#> 
#> $`__default_001`$styleRef
#> character(0)
#> 
#> $`__default_001`$order
#> [1] 999
#> 
#> 
#> attr(,"class")
#> [1] "TFL_options"
#> 
#> $styles
#> $b
#> $b$font
#> $b$font$bold
#> [1] TRUE
#> 
#> 
#> 
#> $i
#> $i$font
#> $i$font$italic
#> [1] TRUE
#> 
#> 
#> 
#> $u
#> $u$font
#> $u$font$underline
#> [1] TRUE
#> 
#> 
#> 
#> $font_bold
#> $font_bold$font
#> $font_bold$font$bold
#> [1] TRUE
#> 
#> 
#> 
#> $font_italic
#> $font_italic$font
#> $font_italic$font$italic
#> [1] TRUE
#> 
#> 
#> 
#> $font_underline
#> $font_underline$font
#> $font_underline$font$underline
#> [1] TRUE
#> 
#> 
#> 
#> $font_arial
#> $font_arial$font
#> $font_arial$font$font_name
#> [1] "Arial"
#> 
#> 
#> 
#> $font_courier_new
#> $font_courier_new$font
#> $font_courier_new$font$font_name
#> [1] "Courier New"
#> 
#> 
#> 
#> $font_times_new_roman
#> $font_times_new_roman$font
#> $font_times_new_roman$font$font_name
#> [1] "Times New Roman"
#> 
#> 
#> 
#> $font_georgia
#> $font_georgia$font
#> $font_georgia$font$font_name
#> [1] "Georgia"
#> 
#> 
#> 
#> $font_verdana
#> $font_verdana$font
#> $font_verdana$font$font_name
#> [1] "Verdana"
#> 
#> 
#> 
#> $font_trebuchet_ms
#> $font_trebuchet_ms$font
#> $font_trebuchet_ms$font$font_name
#> [1] "Trebuchet MS"
#> 
#> 
#> 
#> $fs_7
#> $fs_7$font
#> $fs_7$font$font_size
#> [1] "7pt"
#> 
#> 
#> 
#> $fs_8
#> $fs_8$font
#> $fs_8$font$font_size
#> [1] "8pt"
#> 
#> 
#> 
#> $fs_9
#> $fs_9$font
#> $fs_9$font$font_size
#> [1] "9pt"
#> 
#> 
#> 
#> $fs_10
#> $fs_10$font
#> $fs_10$font$font_size
#> [1] "10pt"
#> 
#> 
#> 
#> $fs_11
#> $fs_11$font
#> $fs_11$font$font_size
#> [1] "11pt"
#> 
#> 
#> 
#> $fs_12
#> $fs_12$font
#> $fs_12$font$font_size
#> [1] "12pt"
#> 
#> 
#> 
#> $fc_black
#> $fc_black$font
#> $fc_black$font$color
#> [1] "#000000"
#> 
#> 
#> 
#> $fc_red
#> $fc_red$font
#> $fc_red$font$color
#> [1] "#FF0000"
#> 
#> 
#> 
#> $fc_blue
#> $fc_blue$font
#> $fc_blue$font$color
#> [1] "#0000FF"
#> 
#> 
#> 
#> $fc_green
#> $fc_green$font
#> $fc_green$font$color
#> [1] "#008000"
#> 
#> 
#> 
#> $fc_gray
#> $fc_gray$font
#> $fc_gray$font$color
#> [1] "#595959"
#> 
#> 
#> 
#> $fc_grey
#> $fc_grey$font
#> $fc_grey$font$color
#> [1] "#595959"
#> 
#> 
#> 
#> $fc_navy
#> $fc_navy$font
#> $fc_navy$font$color
#> [1] "#1F3864"
#> 
#> 
#> 
#> $fc_teal
#> $fc_teal$font
#> $fc_teal$font$color
#> [1] "#2E75B6"
#> 
#> 
#> 
#> $fc_olive
#> $fc_olive$font
#> $fc_olive$font$color
#> [1] "#5C7A29"
#> 
#> 
#> 
#> $fc_rust
#> $fc_rust$font
#> $fc_rust$font$color
#> [1] "#C0392B"
#> 
#> 
#> 
#> $fc_plum
#> $fc_plum$font
#> $fc_plum$font$color
#> [1] "#7B2D8B"
#> 
#> 
#> 
#> $fc_slate
#> $fc_slate$font
#> $fc_slate$font$color
#> [1] "#44546A"
#> 
#> 
#> 
#> $hl_yellow
#> $hl_yellow$font
#> $hl_yellow$font$highlight
#> [1] "#FFFF00"
#> 
#> 
#> 
#> $hl_red
#> $hl_red$font
#> $hl_red$font$highlight
#> [1] "#FF0000"
#> 
#> 
#> 
#> $hl_green
#> $hl_green$font
#> $hl_green$font$highlight
#> [1] "#90EE90"
#> 
#> 
#> 
#> $hl_gray
#> $hl_gray$font
#> $hl_gray$font$highlight
#> [1] "#EEEEEE"
#> 
#> 
#> 
#> $hl_grey
#> $hl_grey$font
#> $hl_grey$font$highlight
#> [1] "#EEEEEE"
#> 
#> 
#> 
#> $hl_peach
#> $hl_peach$font
#> $hl_peach$font$highlight
#> [1] "#FADADD"
#> 
#> 
#> 
#> $hl_mint
#> $hl_mint$font
#> $hl_mint$font$highlight
#> [1] "#D5F5E3"
#> 
#> 
#> 
#> $hl_sky
#> $hl_sky$font
#> $hl_sky$font$highlight
#> [1] "#D6EAF8"
#> 
#> 
#> 
#> $hl_lemon
#> $hl_lemon$font
#> $hl_lemon$font$highlight
#> [1] "#FEFBD8"
#> 
#> 
#> 
#> $hl_lilac
#> $hl_lilac$font
#> $hl_lilac$font$highlight
#> [1] "#E8DAEF"
#> 
#> 
#> 
#> $al
#> $al$paragraph
#> $al$paragraph$alignment
#> [1] "left"
#> 
#> 
#> 
#> $ar
#> $ar$paragraph
#> $ar$paragraph$alignment
#> [1] "right"
#> 
#> 
#> 
#> $ac
#> $ac$paragraph
#> $ac$paragraph$alignment
#> [1] "center"
#> 
#> 
#> 
#> $text_left
#> $text_left$paragraph
#> $text_left$paragraph$alignment
#> [1] "left"
#> 
#> 
#> 
#> $text_right
#> $text_right$paragraph
#> $text_right$paragraph$alignment
#> [1] "right"
#> 
#> 
#> 
#> $text_center
#> $text_center$paragraph
#> $text_center$paragraph$alignment
#> [1] "center"
#> 
#> 
#> 
#> $ind0
#> $ind0$paragraph
#> $ind0$paragraph$indents
#> $ind0$paragraph$indents$left
#> [1] "0cm"
#> 
#> 
#> 
#> 
#> $ind1
#> $ind1$paragraph
#> $ind1$paragraph$indents
#> $ind1$paragraph$indents$left
#> [1] "0.5cm"
#> 
#> 
#> 
#> 
#> $ind2
#> $ind2$paragraph
#> $ind2$paragraph$indents
#> $ind2$paragraph$indents$left
#> [1] "1.0cm"
#> 
#> 
#> 
#> 
#> $ind3
#> $ind3$paragraph
#> $ind3$paragraph$indents
#> $ind3$paragraph$indents$left
#> [1] "1.5cm"
#> 
#> 
#> 
#> 
#> $ind4
#> $ind4$paragraph
#> $ind4$paragraph$indents
#> $ind4$paragraph$indents$left
#> [1] "2.0cm"
#> 
#> 
#> 
#> 
#> $indent_0
#> $indent_0$paragraph
#> $indent_0$paragraph$indents
#> $indent_0$paragraph$indents$left
#> [1] "0cm"
#> 
#> 
#> 
#> 
#> $indent_1
#> $indent_1$paragraph
#> $indent_1$paragraph$indents
#> $indent_1$paragraph$indents$left
#> [1] "0.5cm"
#> 
#> 
#> 
#> 
#> $indent_2
#> $indent_2$paragraph
#> $indent_2$paragraph$indents
#> $indent_2$paragraph$indents$left
#> [1] "1.0cm"
#> 
#> 
#> 
#> 
#> $indent_3
#> $indent_3$paragraph
#> $indent_3$paragraph$indents
#> $indent_3$paragraph$indents$left
#> [1] "1.5cm"
#> 
#> 
#> 
#> 
#> $indent_4
#> $indent_4$paragraph
#> $indent_4$paragraph$indents
#> $indent_4$paragraph$indents$left
#> [1] "2.0cm"
#> 
#> 
#> 
#> 
#> $rind0
#> $rind0$paragraph
#> $rind0$paragraph$indents
#> $rind0$paragraph$indents$right
#> [1] "0cm"
#> 
#> 
#> 
#> 
#> $rind1
#> $rind1$paragraph
#> $rind1$paragraph$indents
#> $rind1$paragraph$indents$right
#> [1] "0.5cm"
#> 
#> 
#> 
#> 
#> $rind2
#> $rind2$paragraph
#> $rind2$paragraph$indents
#> $rind2$paragraph$indents$right
#> [1] "1.0cm"
#> 
#> 
#> 
#> 
#> $rind3
#> $rind3$paragraph
#> $rind3$paragraph$indents
#> $rind3$paragraph$indents$right
#> [1] "1.5cm"
#> 
#> 
#> 
#> 
#> $rind4
#> $rind4$paragraph
#> $rind4$paragraph$indents
#> $rind4$paragraph$indents$right
#> [1] "2.0cm"
#> 
#> 
#> 
#> 
#> $rindent_0
#> $rindent_0$paragraph
#> $rindent_0$paragraph$indents
#> $rindent_0$paragraph$indents$right
#> [1] "0cm"
#> 
#> 
#> 
#> 
#> $rindent_1
#> $rindent_1$paragraph
#> $rindent_1$paragraph$indents
#> $rindent_1$paragraph$indents$right
#> [1] "0.5cm"
#> 
#> 
#> 
#> 
#> $rindent_2
#> $rindent_2$paragraph
#> $rindent_2$paragraph$indents
#> $rindent_2$paragraph$indents$right
#> [1] "1.0cm"
#> 
#> 
#> 
#> 
#> $rindent_3
#> $rindent_3$paragraph
#> $rindent_3$paragraph$indents
#> $rindent_3$paragraph$indents$right
#> [1] "1.5cm"
#> 
#> 
#> 
#> 
#> $rindent_4
#> $rindent_4$paragraph
#> $rindent_4$paragraph$indents
#> $rindent_4$paragraph$indents$right
#> [1] "2.0cm"
#> 
#> 
#> 
#> 
#> $tw_95
#> $tw_95$paragraph
#> $tw_95$paragraph$indents
#> $tw_95$paragraph$indents$left
#> [1] "0.68cm"
#> 
#> $tw_95$paragraph$indents$right
#> [1] "0.68cm"
#> 
#> 
#> 
#> 
#> $tw_90
#> $tw_90$paragraph
#> $tw_90$paragraph$indents
#> $tw_90$paragraph$indents$left
#> [1] "1.36cm"
#> 
#> $tw_90$paragraph$indents$right
#> [1] "1.36cm"
#> 
#> 
#> 
#> 
#> $tw_85
#> $tw_85$paragraph
#> $tw_85$paragraph$indents
#> $tw_85$paragraph$indents$left
#> [1] "2.04cm"
#> 
#> $tw_85$paragraph$indents$right
#> [1] "2.04cm"
#> 
#> 
#> 
#> 
#> $tw_80
#> $tw_80$paragraph
#> $tw_80$paragraph$indents
#> $tw_80$paragraph$indents$left
#> [1] "2.72cm"
#> 
#> $tw_80$paragraph$indents$right
#> [1] "2.72cm"
#> 
#> 
#> 
#> 
#> $tw_75
#> $tw_75$paragraph
#> $tw_75$paragraph$indents
#> $tw_75$paragraph$indents$left
#> [1] "3.40cm"
#> 
#> $tw_75$paragraph$indents$right
#> [1] "3.40cm"
#> 
#> 
#> 
#> 
#> $tw_70
#> $tw_70$paragraph
#> $tw_70$paragraph$indents
#> $tw_70$paragraph$indents$left
#> [1] "4.07cm"
#> 
#> $tw_70$paragraph$indents$right
#> [1] "4.07cm"
#> 
#> 
#> 
#> 
#> $tw_65
#> $tw_65$paragraph
#> $tw_65$paragraph$indents
#> $tw_65$paragraph$indents$left
#> [1] "4.75cm"
#> 
#> $tw_65$paragraph$indents$right
#> [1] "4.75cm"
#> 
#> 
#> 
#> 
#> $tw_60
#> $tw_60$paragraph
#> $tw_60$paragraph$indents
#> $tw_60$paragraph$indents$left
#> [1] "5.43cm"
#> 
#> $tw_60$paragraph$indents$right
#> [1] "5.43cm"
#> 
#> 
#> 
#> 
#> $tw_55
#> $tw_55$paragraph
#> $tw_55$paragraph$indents
#> $tw_55$paragraph$indents$left
#> [1] "6.11cm"
#> 
#> $tw_55$paragraph$indents$right
#> [1] "6.11cm"
#> 
#> 
#> 
#> 
#> $tw_50
#> $tw_50$paragraph
#> $tw_50$paragraph$indents
#> $tw_50$paragraph$indents$left
#> [1] "6.79cm"
#> 
#> $tw_50$paragraph$indents$right
#> [1] "6.79cm"
#> 
#> 
#> 
#> 
#> $sp_0
#> $sp_0$paragraph
#> $sp_0$paragraph$spacing
#> $sp_0$paragraph$spacing$before
#> [1] "0pt"
#> 
#> $sp_0$paragraph$spacing$after
#> [1] "0pt"
#> 
#> $sp_0$paragraph$spacing$line_spacing
#> [1] 1
#> 
#> 
#> 
#> 
#> $sp_2
#> $sp_2$paragraph
#> $sp_2$paragraph$spacing
#> $sp_2$paragraph$spacing$before
#> [1] "2pt"
#> 
#> $sp_2$paragraph$spacing$after
#> [1] "2pt"
#> 
#> $sp_2$paragraph$spacing$line_spacing
#> [1] 1
#> 
#> 
#> 
#> 
#> $sp_4
#> $sp_4$paragraph
#> $sp_4$paragraph$spacing
#> $sp_4$paragraph$spacing$before
#> [1] "4pt"
#> 
#> $sp_4$paragraph$spacing$after
#> [1] "4pt"
#> 
#> $sp_4$paragraph$spacing$line_spacing
#> [1] 1
#> 
#> 
#> 
#> 
#> $kl
#> $kl$paragraph
#> $kl$paragraph$keep_lines
#> [1] TRUE
#> 
#> 
#> 
#> $kn
#> $kn$paragraph
#> $kn$paragraph$keep_next
#> [1] TRUE
#> 
#> 
#> 
#> $grp_hdr
#> $grp_hdr$font
#> $grp_hdr$font$bold
#> [1] TRUE
#> 
#> 
#> $grp_hdr$paragraph
#> $grp_hdr$paragraph$indents
#> $grp_hdr$paragraph$indents$left
#> [1] "0cm"
#> 
#> 
#> $grp_hdr$paragraph$spacing
#> $grp_hdr$paragraph$spacing$before
#> [1] "4pt"
#> 
#> $grp_hdr$paragraph$spacing$after
#> [1] "0pt"
#> 
#> 
#> 
#> 
#> $grp_hdr_i
#> $grp_hdr_i$font
#> $grp_hdr_i$font$bold
#> [1] TRUE
#> 
#> $grp_hdr_i$font$italic
#> [1] TRUE
#> 
#> 
#> $grp_hdr_i$paragraph
#> $grp_hdr_i$paragraph$indents
#> $grp_hdr_i$paragraph$indents$left
#> [1] "0cm"
#> 
#> 
#> $grp_hdr_i$paragraph$spacing
#> $grp_hdr_i$paragraph$spacing$before
#> [1] "4pt"
#> 
#> $grp_hdr_i$paragraph$spacing$after
#> [1] "0pt"
#> 
#> 
#> 
#> 
#> $va_t
#> $va_t$table_style
#> $va_t$table_style$vertical_alignment
#> [1] "top"
#> 
#> 
#> 
#> $va_m
#> $va_m$table_style
#> $va_m$table_style$vertical_alignment
#> [1] "center"
#> 
#> 
#> 
#> $va_b
#> $va_b$table_style
#> $va_b$table_style$vertical_alignment
#> [1] "bottom"
#> 
#> 
#> 
#> $va_top
#> $va_top$table_style
#> $va_top$table_style$vertical_alignment
#> [1] "top"
#> 
#> 
#> 
#> $va_center
#> $va_center$table_style
#> $va_center$table_style$vertical_alignment
#> [1] "center"
#> 
#> 
#> 
#> $va_bottom
#> $va_bottom$table_style
#> $va_bottom$table_style$vertical_alignment
#> [1] "bottom"
#> 
#> 
#> 
#> $to_h
#> $to_h$table_style
#> $to_h$table_style$text_orientation
#> [1] "horizontal"
#> 
#> 
#> 
#> $to_90
#> $to_90$table_style
#> $to_90$table_style$text_orientation
#> [1] "vertical_90"
#> 
#> 
#> 
#> $to_270
#> $to_270$table_style
#> $to_270$table_style$text_orientation
#> [1] "vertical_270"
#> 
#> 
#> 
#> $text_horizontal
#> $text_horizontal$table_style
#> $text_horizontal$table_style$text_orientation
#> [1] "horizontal"
#> 
#> 
#> 
#> $text_vertical_90
#> $text_vertical_90$table_style
#> $text_vertical_90$table_style$text_orientation
#> [1] "vertical_90"
#> 
#> 
#> 
#> $text_vertical_270
#> $text_vertical_270$table_style
#> $text_vertical_270$table_style$text_orientation
#> [1] "vertical_270"
#> 
#> 
#> 
#> $bg_blue
#> $bg_blue$table_style
#> $bg_blue$table_style$background_color
#> [1] "#D9E1F2"
#> 
#> 
#> 
#> $bg_gray
#> $bg_gray$table_style
#> $bg_gray$table_style$background_color
#> [1] "#F2F2F2"
#> 
#> 
#> 
#> $bg_grey
#> $bg_grey$table_style
#> $bg_grey$table_style$background_color
#> [1] "#F2F2F2"
#> 
#> 
#> 
#> $bg_peach
#> $bg_peach$table_style
#> $bg_peach$table_style$background_color
#> [1] "#FADADD"
#> 
#> 
#> 
#> $bg_mint
#> $bg_mint$table_style
#> $bg_mint$table_style$background_color
#> [1] "#D5F5E3"
#> 
#> 
#> 
#> $bg_sky
#> $bg_sky$table_style
#> $bg_sky$table_style$background_color
#> [1] "#D6EAF8"
#> 
#> 
#> 
#> $bg_lemon
#> $bg_lemon$table_style
#> $bg_lemon$table_style$background_color
#> [1] "#FEFBD8"
#> 
#> 
#> 
#> $bg_lilac
#> $bg_lilac$table_style
#> $bg_lilac$table_style$background_color
#> [1] "#E8DAEF"
#> 
#> 
#> 
#> $bg_navy
#> $bg_navy$table_style
#> $bg_navy$table_style$background_color
#> [1] "#1F3864"
#> 
#> 
#> 
#> $bg_slate
#> $bg_slate$table_style
#> $bg_slate$table_style$background_color
#> [1] "#44546A"
#> 
#> 
#> 
#> $bg_steel
#> $bg_steel$table_style
#> $bg_steel$table_style$background_color
#> [1] "#BDD7EE"
#> 
#> 
#> 
#> $row_h2
#> $row_h2$table_style
#> $row_h2$table_style$row_height
#> [1] "2pt"
#> 
#> 
#> 
#> $row_h4
#> $row_h4$table_style
#> $row_h4$table_style$row_height
#> [1] "4pt"
#> 
#> 
#> 
#> $row_h6
#> $row_h6$table_style
#> $row_h6$table_style$row_height
#> [1] "6pt"
#> 
#> 
#> 
#> $bt
#> $bt$table_style
#> $bt$table_style$borders
#> $bt$table_style$borders$top
#> $bt$table_style$borders$top$width
#> [1] "1pt"
#> 
#> $bt$table_style$borders$top$line_style
#> [1] "single"
#> 
#> $bt$table_style$borders$top$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> $bb
#> $bb$table_style
#> $bb$table_style$borders
#> $bb$table_style$borders$bottom
#> $bb$table_style$borders$bottom$width
#> [1] "1pt"
#> 
#> $bb$table_style$borders$bottom$line_style
#> [1] "single"
#> 
#> $bb$table_style$borders$bottom$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> $bl
#> $bl$table_style
#> $bl$table_style$borders
#> $bl$table_style$borders$left
#> $bl$table_style$borders$left$width
#> [1] "1pt"
#> 
#> $bl$table_style$borders$left$line_style
#> [1] "single"
#> 
#> $bl$table_style$borders$left$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> $br
#> $br$table_style
#> $br$table_style$borders
#> $br$table_style$borders$right
#> $br$table_style$borders$right$width
#> [1] "1pt"
#> 
#> $br$table_style$borders$right$line_style
#> [1] "single"
#> 
#> $br$table_style$borders$right$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> $bt_th
#> $bt_th$table_style
#> $bt_th$table_style$borders
#> $bt_th$table_style$borders$top
#> $bt_th$table_style$borders$top$width
#> [1] "0.5pt"
#> 
#> $bt_th$table_style$borders$top$line_style
#> [1] "single"
#> 
#> $bt_th$table_style$borders$top$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> $bb_th
#> $bb_th$table_style
#> $bb_th$table_style$borders
#> $bb_th$table_style$borders$bottom
#> $bb_th$table_style$borders$bottom$width
#> [1] "0.5pt"
#> 
#> $bb_th$table_style$borders$bottom$line_style
#> [1] "single"
#> 
#> $bb_th$table_style$borders$bottom$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> $bc_gray
#> $bc_gray$table_style
#> $bc_gray$table_style$borders
#> $bc_gray$table_style$borders$top
#> $bc_gray$table_style$borders$top$color
#> [1] "#AAAAAA"
#> 
#> 
#> $bc_gray$table_style$borders$bottom
#> $bc_gray$table_style$borders$bottom$color
#> [1] "#AAAAAA"
#> 
#> 
#> $bc_gray$table_style$borders$left
#> $bc_gray$table_style$borders$left$color
#> [1] "#AAAAAA"
#> 
#> 
#> $bc_gray$table_style$borders$right
#> $bc_gray$table_style$borders$right$color
#> [1] "#AAAAAA"
#> 
#> 
#> 
#> 
#> 
#> $bc_grey
#> $bc_grey$table_style
#> $bc_grey$table_style$borders
#> $bc_grey$table_style$borders$top
#> $bc_grey$table_style$borders$top$color
#> [1] "#AAAAAA"
#> 
#> 
#> $bc_grey$table_style$borders$bottom
#> $bc_grey$table_style$borders$bottom$color
#> [1] "#AAAAAA"
#> 
#> 
#> $bc_grey$table_style$borders$left
#> $bc_grey$table_style$borders$left$color
#> [1] "#AAAAAA"
#> 
#> 
#> $bc_grey$table_style$borders$right
#> $bc_grey$table_style$borders$right$color
#> [1] "#AAAAAA"
#> 
#> 
#> 
#> 
#> 
#> $bc_white
#> $bc_white$table_style
#> $bc_white$table_style$borders
#> $bc_white$table_style$borders$top
#> $bc_white$table_style$borders$top$width
#> [1] "0pt"
#> 
#> $bc_white$table_style$borders$top$line_style
#> [1] "none"
#> 
#> $bc_white$table_style$borders$top$color
#> [1] "#FFFFFF"
#> 
#> 
#> $bc_white$table_style$borders$bottom
#> $bc_white$table_style$borders$bottom$width
#> [1] "0pt"
#> 
#> $bc_white$table_style$borders$bottom$line_style
#> [1] "none"
#> 
#> $bc_white$table_style$borders$bottom$color
#> [1] "#FFFFFF"
#> 
#> 
#> $bc_white$table_style$borders$left
#> $bc_white$table_style$borders$left$width
#> [1] "0pt"
#> 
#> $bc_white$table_style$borders$left$line_style
#> [1] "none"
#> 
#> $bc_white$table_style$borders$left$color
#> [1] "#FFFFFF"
#> 
#> 
#> $bc_white$table_style$borders$right
#> $bc_white$table_style$borders$right$width
#> [1] "0pt"
#> 
#> $bc_white$table_style$borders$right$line_style
#> [1] "none"
#> 
#> $bc_white$table_style$borders$right$color
#> [1] "#FFFFFF"
#> 
#> 
#> 
#> 
#> 
#> $brw_thick
#> $brw_thick$table_style
#> $brw_thick$table_style$borders
#> $brw_thick$table_style$borders$right
#> $brw_thick$table_style$borders$right$width
#> [1] "4pt"
#> 
#> $brw_thick$table_style$borders$right$line_style
#> [1] "single"
#> 
#> $brw_thick$table_style$borders$right$color
#> [1] "#FFFFFF"
#> 
#> 
#> 
#> 
#> 
#> $blw_thick
#> $blw_thick$table_style
#> $blw_thick$table_style$borders
#> $blw_thick$table_style$borders$left
#> $blw_thick$table_style$borders$left$width
#> [1] "4pt"
#> 
#> $blw_thick$table_style$borders$left$line_style
#> [1] "single"
#> 
#> $blw_thick$table_style$borders$left$color
#> [1] "#FFFFFF"
#> 
#> 
#> 
#> 
#> 
#> $pb
#> $pb$paragraph
#> $pb$paragraph$borders
#> $pb$paragraph$borders$bottom
#> $pb$paragraph$borders$bottom$width
#> [1] "1pt"
#> 
#> $pb$paragraph$borders$bottom$line_style
#> [1] "single"
#> 
#> $pb$paragraph$borders$bottom$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> $pb_th
#> $pb_th$paragraph
#> $pb_th$paragraph$borders
#> $pb_th$paragraph$borders$bottom
#> $pb_th$paragraph$borders$bottom$width
#> [1] "0.5pt"
#> 
#> $pb_th$paragraph$borders$bottom$line_style
#> [1] "single"
#> 
#> $pb_th$paragraph$borders$bottom$color
#> [1] "#000000"
#> 
#> 
#> 
#> 
#> 
#> attr(,"class")
#> [1] "TFL_options"
#> 
#> $insertTOC
#> [1] FALSE
#> 
#> $tocTitle
#> [1] "Table of Contents"
#> 
#> $output_directory
#> [1] "."
#> 
#> $meta_directory
#> NULL
#> 
#> attr(,"class")
#> [1] "TFL_options"